pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bson>=0.5.10",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.13.2",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
]
